fix(api): cache get preview link db lookup
